ajuste organizacao
Repassados testes referentes a stock e ticket para seus respectivos arquivos para melhor organizacao
